initial action for proper folder structure; relocated code to proper files and folders
