--- test-html/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>Clique</title>
-	</head>
-	<body>
-		<ul id="messages"></ul>
-		<form id="form" action="">
-			<input id="input" autocomplete="off" /><button>Send</button>
-		</form>
-		<script src="/socket.io/socket.io.js"></script>
-		<script>
-			var socket = io();
-		</script>
-		<script src="/socket.io/socket.io.js"></script>
-		<script>
-			var socket = io();
-		</script>
-	</body>
-</html>
